feat: decomposite components
