--- src/app 2/about/layout.tsx
@@ -1,9 +1,0 @@
-import type { PropsWithChildren } from 'react';
-
-export default function Layout({ children }: PropsWithChildren) {
-  return (
-    <>
-      <div>{children}</div>
-    </>
-  );
-}
