--- src/app 2/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default function HomePage() {
-  redirect('/pokemons');
-}
