+++ src/components/forms/uncontrolled/Uncontrolled.module.css
@@ -36,48 +36,3 @@
     background-color: white;
     color: black;
 }
-
-.item {
-    display: flex;
-    flex-direction: column  ;
-    justify-content: space-between;
-    gap: 0px;
-}
-
-.item > input {
-    height: 40px;
-    border: none;
-    border-bottom: 1px solid rgb(198, 197, 197);
-}
-
-.country {
-    position: relative;
-}
-
-.file {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    align-items: center;
-}
-
-.file > img {
-    width: 120px;
-    height: 120px;;
-}
-
-.file > input {
-    display: none;
-}
-
-.file > label {
-    background-color: rgb(52, 137, 227);
-    color: white;
-    padding: 10px;
-    border-radius: 10px;
-    cursor: pointer;
-}
-
-.file > label:hover {
-    background-color: rgb(96, 168, 245);
-}
