+++ src/components/forms/uncontrolled/inputs/agreement/AgreementInput.tsx
@@ -1,0 +1,10 @@
+function AgreementInput() {
+  return (
+    <div>
+      <input type="checkbox" name="agreement" id="agreement" />
+      <label htmlFor="agreement">I agree with Terms and Conditions</label>
+    </div>
+  );
+}
+
+export default AgreementInput;
