+++ src/components/forms/uncontrolled/inputs/avatar/AvatarInput.module.css
@@ -1,0 +1,27 @@
+.file {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    align-items: center;
+}
+
+.file > img {
+    width: 120px;
+    height: 120px;;
+}
+
+.file > input {
+    display: none;
+}
+
+.file > label {
+    background-color: rgb(52, 137, 227);
+    color: white;
+    padding: 10px;
+    border-radius: 10px;
+    cursor: pointer;
+}
+
+.file > label:hover {
+    background-color: rgb(96, 168, 245);
+}
