+++ src/components/forms/uncontrolled/inputs/confirm/ConfirmInput.tsx
@@ -1,0 +1,17 @@
+import styles from '../Inputs.module.css';
+
+function ConfirmInput() {
+  return (
+    <div className={styles.item}>
+      <label htmlFor="confirm">Confirm password</label>
+      <input
+        type="text"
+        name="confirm"
+        id="confirm"
+        placeholder="Your password"
+      />
+    </div>
+  );
+}
+
+export default ConfirmInput;
