+++ src/components/forms/uncontrolled/inputs/email/EmailInput.tsx
@@ -1,0 +1,12 @@
+import styles from '../Inputs.module.css';
+
+function EmailInput() {
+  return (
+    <div className={styles.item}>
+      <label htmlFor="email">Email</label>
+      <input type="text" name="email" id="email" placeholder="your@email.com" />
+    </div>
+  );
+}
+
+export default EmailInput;
