+++ src/components/forms/uncontrolled/inputs/name/NameInput.tsx
@@ -1,0 +1,12 @@
+import styles from '../Inputs.module.css';
+
+function NameInput() {
+  return (
+    <div className={styles.item}>
+      <label htmlFor="name">Name</label>
+      <input type="text" name="name" id="name" placeholder="Your name" />
+    </div>
+  );
+}
+
+export default NameInput;
